fix: transfers professionals' data to external files, simulating API request
